+++ rpms/SPECS/3.7.0/wazuh-manager-3.7.0.spec
@@ -235,7 +235,11 @@ if [ $1 = 1 ]; then
     sles=$(grep "\"sles" /etc/os-release)
     if [ ! -z "$sles" ]; then
       if [ -f /etc/init.d/init.d/wazuh-manager ]; then
-        rm -rf /etc/init.d/init.d/
+        rm -f /etc/init.d/init.d/wazuh-manager
+        # Delete the directory if it is empty
+        if [ -z "$(ls -A /etc/init.d/init.d/)" ]; then
+          rm -rf /etc/init.d/init.d/
+        fi
       fi
       install -m 755 %{_localstatedir}/ossec/tmp/src/init/ossec-hids-suse.init /etc/init.d/wazuh-manager
     fi
